check extension enablement at activation time and on command invokation
